// FootballManagement/football-ng/src/app/Model/player-in-team-tournament.ts
export class PlayerInTeamTournament {
  id: number;
  teamTourId: number;
  teamName: string;
  tournamentName: string;
  playerId: number;
  playerName: string;
  nbClothes: number;
  nationality: string;
  totalRedCards: number;
  totalYellowCards: number;
  totalGoals: number;
  totalMatchesJoined: number;
  dob: Date;
}